#7 - Adicionada lógica de filtragem de projetos por teste, fonte, e pacote, e melhorada a utilização de diversos prompts.
